requirements.txt
langchain==1.2.13
langchain-openai==1.1.1
langgraph==1.1.3
python-dotenv==1.2.2
